Move Settings to its own module.
