+++ overwatch/src/services/settings/mod.rs
@@ -1,0 +1,7 @@
+pub mod handle;
+pub mod notifier;
+pub mod updater;
+
+pub use handle::SettingsHandle;
+pub use notifier::SettingsNotifier;
+pub use updater::SettingsUpdater;
